flask-server/requirements.txt
click==8.1.2
Flask==2.1.1
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
Werkzeug==2.1.1
zipp==3.8.0
